+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add commands to export
 - Add tests for get ref and set ref
 - Add get version command
+- Add tests for process dependencies
 
 ### Changed